ubah struktur folder

+++ routes/web.php
@@ -50,51 +50,51 @@ Route::middleware('auth')->group(function () {
 });
 
 Route::middleware(['auth', 'permission:request_kelas'])->group(function () {
-    Route::get('/kelas/request', [KelasController::class, 'requestKelas'])->name('kelas.request');
+    Route::get('admin/kelas/request', [KelasController::class, 'requestKelas'])->name('kelas.request');
 });
 
 Route::middleware(['auth', 'permission:lihat_kelas'])->group(function () {
-    Route::get('/kelas', [KelasController::class, 'index'])->name('kelas.index');
+    Route::get('admin/kelas', [KelasController::class, 'index'])->name('kelas.index');
 });
 
 Route::middleware(['auth', 'permission:lihat_member_kelas'])->group(function () {
-    Route::get('/kelas/{id}/member', [KelasController::class, 'memberList'])->name('kelas.member');
+    Route::get('admin/kelas/{id}/member', [KelasController::class, 'memberList'])->name('kelas.member');
 });
 
 Route::middleware(['auth', 'permission:buat_kelas'])->group(function () {
-    Route::get('/kelas/create', [KelasController::class, 'create'])->name('kelas.create');
-    Route::post('/kelas', [KelasController::class, 'store'])->name('kelas.store');
+    Route::get('admin/kelas/create', [KelasController::class, 'create'])->name('kelas.create');
+    Route::post('admin/kelas', [KelasController::class, 'store'])->name('kelas.store');
 });
 
 Route::middleware(['auth', 'permission:ubah_kelas'])->group(function () {
-    Route::get('/kelas/{id}/edit', [KelasController::class, 'edit'])->name('kelas.edit');
-    Route::put('/kelas/{id}', [KelasController::class, 'update'])->name('kelas.update');
+    Route::get('admin/kelas/{id}/edit', [KelasController::class, 'edit'])->name('kelas.edit');
+    Route::put('admin/kelas/{id}', [KelasController::class, 'update'])->name('kelas.update');
 });
 
 Route::middleware(['auth', 'permission:hapus_kelas'])->group(function () {
-    Route::delete('/kelas/{id}', [KelasController::class, 'destroy'])->name('kelas.destroy');
+    Route::delete('members/kelas/{id}', [KelasController::class, 'destroy'])->name('kelas.destroy');
 });
 
 Route::middleware(['auth', 'permission:lihat_member_membership'])->group(function () {
 });
 
 Route::middleware(['auth', 'permission:tambah_membership'])->group(function () {
-    Route::get('/membership/create', [MembershipController::class, 'create'])->name('membership.create');
-    Route::post('/membership', [MembershipController::class, 'store'])->name('membership.store');
+    Route::get('admin/membership/create', [MembershipController::class, 'create'])->name('membership.create');
+    Route::post('admin/membership', [MembershipController::class, 'store'])->name('membership.store');
 });
 
 Route::middleware(['auth', 'permission:ubah_membership'])->group(function () {
-    Route::get('/membership/{id}/edit', [MembershipController::class, 'edit'])->name('membership.edit');
-    Route::put('/membership/{id}', [MembershipController::class, 'update'])->name('membership.update');
+    Route::get('admin/membership/{id}/edit', [MembershipController::class, 'edit'])->name('membership.edit');
+    Route::put('admin/membership/{id}', [MembershipController::class, 'update'])->name('membership.update');
 });
 
 Route::middleware(['auth', 'permission:hapus_membership'])->group(function () {
-    Route::delete('/membership/{id}', [MembershipController::class, 'destroy'])->name('membership.destroy');
+    Route::delete('admin/membership/{id}', [MembershipController::class, 'destroy'])->name('membership.destroy');
 });
 
 Route::middleware(['auth', 'permission:lihat_daftar_membership'])->group(function () {
-    Route::get('/membership', [MembershipController::class, 'index'])->name('membership.index');
-    Route::get('/membership/{id}', [MembershipController::class, 'show'])->name('membership.show');
+    Route::get('admin/membership', [MembershipController::class, 'index'])->name('membership.index');
+    Route::get('admin/membership/{id}', [MembershipController::class, 'show'])->name('membership.show');
 });
 
 
@@ -122,26 +122,26 @@ Route::middleware(['auth', 'permission:ubah_alat'])->group(function () {
 Route::middleware(['auth', 'permission:hapus_alat'])->group(function () {
 });
 Route::middleware(['auth', 'permission:lihat_akun_member'])->group(function () {
-    Route::get('/member', [MemberController::class, 'index'])->name('member.index');
+    Route::get('admin/member', [MemberController::class, 'index'])->name('member.index');
 });
 
 Route::middleware(['auth', 'permission:ubah_akun_member'])->group(function () {
-    Route::get('/member/{id}/edit', [MemberController::class, 'edit'])->name('member.edit');
-    Route::put('/member/{id}', [MemberController::class, 'update'])->name('member.update');
+    Route::get('admin/member/{id}/edit', [MemberController::class, 'edit'])->name('member.edit');
+    Route::put('admin/member/{id}', [MemberController::class, 'update'])->name('member.update');
 });
 
 Route::middleware(['auth', 'permission:hapus_akun_member'])->group(function () {
-    Route::delete('/member/{id}', [MemberController::class, 'destroy'])->name('member.destroy');
+    Route::delete('admin/member/{id}', [MemberController::class, 'destroy'])->name('member.destroy');
 });
 Route::middleware(['auth', 'permission:lihat_akun_pelatih'])->group(function () {
-    Route::get('/pelatih', [PelatihController::class, 'index'])->name('pelatih.index');
+    Route::get('admin/pelatih', [PelatihController::class, 'index'])->name('pelatih.index');
 });
 Route::middleware(['auth', 'permission:ubah_akun_pelatih'])->group(function () {
-    Route::get('/pelatih/{id}/edit', [PelatihController::class, 'edit'])->name('pelatih.edit');
-    Route::put('/pelatih/{id}', [PelatihController::class, 'update'])->name('pelatih.update');
+    Route::get('admin/pelatih/{id}/edit', [PelatihController::class, 'edit'])->name('pelatih.edit');
+    Route::put('admin/pelatih/{id}', [PelatihController::class, 'update'])->name('pelatih.update');
 });
 Route::middleware(['auth', 'permission:hapus_akun_pelatih'])->group(function () {
-    Route::delete('/pelatih/{id}', [PelatihController::class, 'destroy'])->name('pelatih.destroy');
+    Route::delete('admin/pelatih/{id}', [PelatihController::class, 'destroy'])->name('pelatih.destroy');
 });
 
 require __DIR__.'/auth.php';
